day_07_light_rag/requirements.txt
# ── Core ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
loguru==0.7.3

# ── Embeddings (lightweight — no LangChain) ──────────
sentence-transformers>=3.4.1
numpy>=1.26.4

# ── BM25 keyword search ─────────────────────────────
rank-bm25>=0.2.2

# ── Optional LLM generation ─────────────────────────
openai>=1.59.9

# ── Demo & utilities ─────────────────────────────────
streamlit>=1.41.1
requests>=2.32.3
httpx>=0.28.1

# ── Testing ──────────────────────────────────────────
pytest>=8.3.4
